+++ src/generators/generator.ts
@@ -126,6 +126,9 @@ export class Generator {
             throw new Error(`Invalid URI: ${uri}`);
         }
 
+        // console log the size of the data we're saving in case we blow a limit
+        console.log(`Saving ${this.dataType} data of size ${this.data.length}`);
+
         const authHeader = await signedAuthHeader(this.email);
         const response = await fetch(this.resourceUri, {
             method: 'PUT',
@@ -138,7 +141,7 @@ export class Generator {
 
         if (!response.ok) {
             const errorText = await response.text() || 'Unknown Error';
-            throw new Error(`Unable to Save Generated Resource: ${response.status} - ${errorText}`);
+            throw new Error(`Unable to Save Generated Resource (size:${this.data.length} bytes): ${response.status} - ${errorText}`);
         }
 
         console.log(`Saved ${this.dataType} data`);